move AddressSearchProps into types.d
