+++ app/javascript/packages/address-search/components/address-search.tsx
@@ -1,20 +1,9 @@
-import { ComponentType, Dispatch, SetStateAction, useState } from 'react';
+import { useState } from 'react';
 import { Alert, PageHeading } from '@18f/identity-components';
 import { t } from '@18f/identity-i18n';
-import type { RegisterFieldCallback } from '@18f/identity-form-steps';
 import InPersonLocations from './in-person-locations';
 import AddressInput from './address-input';
-import type { LocationQuery, FormattedLocation } from '../types';
-
-interface AddressSearchProps {
-  addressSearchURL: string;
-  disabled: boolean;
-  handleLocationSelect: ((e: any, id: number) => Promise<void>) | null | undefined;
-  resultsHeaderComponent?: ComponentType;
-  locationsURL: string;
-  onFoundLocations: Dispatch<SetStateAction<FormattedLocation[] | null | undefined>>;
-  registerField: RegisterFieldCallback;
-}
+import type { AddressSearchProps, LocationQuery, FormattedLocation } from '../types';
 
 function AddressSearch({
   addressSearchURL,
